replace event model
