+++ adaptive_hockey_federation/main/models.py
@@ -1,7 +1,6 @@
-from datetime import timedelta
-
 from django.db import models
 from django.utils.translation import gettext_lazy as _
+from events.models import Event
 from users.models import User
 
 CHAR_FIELD_LENGTH = 256
@@ -310,7 +309,7 @@ class Team(BaseUniqueName):
         help_text=_('Куратор команды')
     )
     events = models.ManyToManyField(
-        'Event',
+        Event,
         related_name='events',
         verbose_name=_('Соревнования в которых участвует команда'),
         help_text=_('Соревнования в которых участвует команда'),
@@ -439,44 +438,6 @@ class Player(BasePerson):
 
     def __str__(self):
         return ' '.join([self.surname, self.name, self.patronymic])
-
-
-class Event(models.Model):
-    """
-    Модель соревнований.
-    """
-    title = models.CharField(max_length=100)
-    date_start = models.DateField()
-    date_end = models.DateField()
-    duration = models.DurationField()
-    city = models.ForeignKey(
-        City,
-        on_delete=models.CASCADE,
-        verbose_name=_('Город проведения соревнований'),
-        help_text=_('Город проведения соревнований')
-    )
-    location = models.CharField(max_length=100)
-    teams = models.ManyToManyField(
-        Team,
-        related_name='teams',
-        verbose_name=_('Состав команд участников'),
-        help_text=_('Состав команд участников'),
-        through='TeamInEvent'
-    )
-
-    is_active = models.BooleanField(default=True)
-
-    class Meta:
-        verbose_name = 'Соревнование'
-        verbose_name_plural = 'Соревнования'
-        ordering = ('-date_start',)
-
-    def __str__(self):
-        return self.title
-
-    @property
-    def get_duration(self):
-        return timedelta(self.date_end - self.date_start)
 
 
 class TeamInEvent(models.Model):
